Extract local operation response mapping
